requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-dotenv>=1.0.0
httpx>=0.26.0
authlib>=1.3.0
itsdangerous>=2.1.0
python-multipart>=0.0.6
jinja2>=3.1.0
markdown>=3.5.0
google-api-python-client>=2.116.0
google-auth>=2.27.0
resend>=0.7.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
apscheduler>=3.10.0
